requirements.txt
discord.py
aiofiles
judge0api
